move landingpage and registerpage component folders into parent HomePage
