+++ src/components/HomePage/LandingPage/LandingPage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {connect} from 'react-redux';
 import {Redirect} from 'react-router-dom';
-import LoginForm from '../LoginForm/LoginForm';
+import LoginForm from './LoginForm/LoginForm';
 
 export function LandingPage(props) {
     if (props.loggedIn) {
