+++ src/components/HomePage/LandingPage/LoginForm/LoginForm.js
@@ -1,0 +1,62 @@
+import React from 'react';
+import {Field, reduxForm, focus} from 'redux-form';
+import Input from '../../../Input';
+import {login} from '../../../../actions/auth';
+import {required, nonEmpty} from '../../../../validators';
+import './LoginForm.css'
+
+export class LoginForm extends React.Component {
+  onSubmit(values) {
+    return this.props.dispatch(login(values.username, values.password));
+  }
+
+  render() {
+    let error;
+    if (this.props.error) {
+      error = (
+        <div className="form-error" aria-live="polite">
+          {this.props.error}
+        </div>
+      );
+    }
+    return (
+      <form
+        className="LoginForm"
+        onSubmit={this.props.handleSubmit(values =>
+            this.onSubmit(values)
+        )}>
+        {error}
+        <label htmlFor="username">Username</label>
+        <Field
+          component={Input}
+          type="text"
+          name="username"
+          id="username"
+          validate={[required, nonEmpty]}
+          placeholder="username"
+          className="username"
+          required
+        />
+        <label htmlFor="password">Password</label>
+        <Field
+          component={Input}
+          type="password"
+          name="password"
+          id="password"
+          validate={[required, nonEmpty]}
+          placeholder="password"
+          className="user-password"
+          required
+        />
+        <button className="login-button" disabled={this.props.pristine || this.props.submitting}>
+          Log in
+        </button>
+      </form>
+    );
+  }
+}
+
+export default reduxForm({
+  form: 'login',
+  onSubmitFail: (errors, dispatch) => dispatch(focus('login', 'username'))
+})(LoginForm);
