--- src/components/RegistrationPage/RegistrationPage.js
@@ -1,23 +1,0 @@
-import React from 'react';
-import {connect} from 'react-redux';
-import {Link, Redirect} from 'react-router-dom';
-import RegistrationForm from '../RegistrationForm/RegistrationForm';
-
-function RegistrationPage(props) {
-    if (props.loggedIn) {
-        return <Redirect to="/dashboard" />;
-    }
-    return (
-        <div className="home">
-          <h2 className="access_form">Register</h2>
-          <RegistrationForm />
-          <Link to="/">Login</Link>
-        </div>
-    );
-}
-
-const mapStateToProps = state => ({
-    loggedIn: state.auth.currentUser !== null
-});
-
-export default connect(mapStateToProps)(RegistrationPage);
