--- src/components/ScrollButton/ScrollButton.css
@@ -1,50 +1,0 @@
-html, body{
-  height: 100%;
-}
-
-a {
-  color: white;
-}
-
-.long {
-  width: 100%;
-  background-color: #010e19;
-}
-
-.container {
-  height: 400px;
-}
-
-.scroll {
-  background-color: #003366;
-  width: 30px;
-  height: 30px;
-  position: fixed;
-  bottom: 10px;
-  right: 10px;
-  border-radius: 5px;
-  border: none;
-  cursor: pointer;
-
-  &:hover {
-    opacity: 1;
-  }
-}
-
-.arrow-up {
-  color: white;
-  font-size: 24px;
-  transform: rotate(-135deg);
-  -webkit-transform: rotate(-135deg);
-}
-
-@media only screen and (min-width: 700px) {
-  .scroll {
-    width: 40px;
-    height: 40px;
-  }
-
-  .arrow-up {
-    font-size: 30px;
-  }
-}
